+++ modules/dxp/apps/commerce/commerce-product-service/src/main/java/com/liferay/commerce/product/service/impl/CPDefinitionOptionRelServiceImpl.java
@@ -65,7 +65,7 @@ public class CPDefinitionOptionRelServiceImpl
 	public CPDefinitionOptionRel addCPDefinitionOptionRel(
 			long cpDefinitionId, long cpOptionId, String name,
 			Map<Locale, String> titleMap, Map<Locale, String> descriptionMap,
-			String ddmFormFieldTypeName, int priority, boolean facetable,
+			String ddmFormFieldTypeName, double priority, boolean facetable,
 			boolean required, boolean skuContributor, boolean importOptionValue,
 			ServiceContext serviceContext)
 		throws PortalException {
@@ -247,7 +247,7 @@ public class CPDefinitionOptionRelServiceImpl
 	public CPDefinitionOptionRel updateCPDefinitionOptionRel(
 			long cpDefinitionOptionRelId, long cpOptionId, String name,
 			Map<Locale, String> titleMap, Map<Locale, String> descriptionMap,
-			String ddmFormFieldTypeName, int priority, boolean facetable,
+			String ddmFormFieldTypeName, double priority, boolean facetable,
 			boolean required, boolean skuContributor,
 			ServiceContext serviceContext)
 		throws PortalException {
