+++ modules/dxp/apps/commerce/commerce-product-service/src/main/java/com/liferay/commerce/product/service/impl/CPDefinitionOptionValueRelServiceImpl.java
@@ -39,7 +39,7 @@ public class CPDefinitionOptionValueRelServiceImpl
 	@Override
 	public CPDefinitionOptionValueRel addCPDefinitionOptionValueRel(
 			long cpDefinitionOptionRelId, String name,
-			Map<Locale, String> titleMap, int priority,
+			Map<Locale, String> titleMap, double priority,
 			ServiceContext serviceContext)
 		throws PortalException {
 
@@ -167,7 +167,7 @@ public class CPDefinitionOptionValueRelServiceImpl
 	@Override
 	public CPDefinitionOptionValueRel updateCPDefinitionOptionValueRel(
 			long cpDefinitionOptionValueRelId, String name,
-			Map<Locale, String> titleMap, int priority,
+			Map<Locale, String> titleMap, double priority,
 			ServiceContext serviceContext)
 		throws PortalException {
