+++ modules/dxp/apps/commerce/commerce-product-service/src/main/java/com/liferay/commerce/product/service/impl/CPOptionCategoryServiceImpl.java
@@ -38,7 +38,7 @@ public class CPOptionCategoryServiceImpl
 	@Override
 	public CPOptionCategory addCPOptionCategory(
 			String name, Map<Locale, String> titleMap,
-			Map<Locale, String> descriptionMap, int priority,
+			Map<Locale, String> descriptionMap, double priority,
 			ServiceContext serviceContext)
 		throws PortalException {
 
@@ -125,7 +125,7 @@ public class CPOptionCategoryServiceImpl
 	@Override
 	public CPOptionCategory updateCPOptionCategory(
 			long cpOptionCategoryId, String name, Map<Locale, String> titleMap,
-			Map<Locale, String> descriptionMap, int priority,
+			Map<Locale, String> descriptionMap, double priority,
 			ServiceContext serviceContext)
 		throws PortalException {
